local-server/requirements.txt
# Local Server Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
websockets>=12.0
livekit>=0.15.0
livekit-plugins-openai>=0.6.0
aiofiles>=23.0.0